STM32F103RCTX: allow vector table offset to be overriden on the command line
The hardcoded VECT_TAB_OFFSET has been wrapped in an `#ifndef`, allowing us to optionally define it on the command line.

Similarly, in the linker script, a new symbol is added to calculate the flash origin, allowing us to optionally override it on the command line (`-Wl,--defsym=...).

This is useful to support boards such as the SKR Mini E3 V2.0, which comes with a boot loader in the first 28KB of the flash that expects the firmware vector table offset to be at 0x8007000.